+++ server/src/test/java/org/opensearch/index/remote/RemoteStoreEnumsTests.java
@@ -25,8 +25,7 @@ import static org.opensearch.index.remote.RemoteStoreEnums.DataCategory.TRANSLOG
 import static org.opensearch.index.remote.RemoteStoreEnums.DataType.DATA;
 import static org.opensearch.index.remote.RemoteStoreEnums.DataType.LOCK_FILES;
 import static org.opensearch.index.remote.RemoteStoreEnums.DataType.METADATA;
-import static org.opensearch.index.remote.RemoteStoreEnums.PathHashAlgorithm.FNV_1A_BASE64;
-import static org.opensearch.index.remote.RemoteStoreEnums.PathHashAlgorithm.FNV_1A_COMPOSITE_1;
+import static org.opensearch.index.remote.RemoteStoreEnums.PathHashAlgorithm.FNV_1A;
 import static org.opensearch.index.remote.RemoteStoreEnums.PathType.FIXED;
 import static org.opensearch.index.remote.RemoteStoreEnums.PathType.HASHED_INFIX;
 import static org.opensearch.index.remote.RemoteStoreEnums.PathType.HASHED_PREFIX;
@@ -162,10 +161,10 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        BlobPath result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        BlobPath result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertTrue(
             result.buildAsString()
-                .startsWith(String.join(SEPARATOR, FNV_1A_BASE64.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
+                .startsWith(String.join(SEPARATOR, FNV_1A.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
         );
 
         // assert with exact value for known base path
@@ -179,7 +178,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertEquals("DgSI70IciXs/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/translog/data/", result.buildAsString());
 
         // Translog Metadata
@@ -191,10 +190,10 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertTrue(
             result.buildAsString()
-                .startsWith(String.join(SEPARATOR, FNV_1A_BASE64.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
+                .startsWith(String.join(SEPARATOR, FNV_1A.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
         );
 
         // assert with exact value for known base path
@@ -205,7 +204,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertEquals("oKU5SjILiy4/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/translog/metadata/", result.buildAsString());
 
         // Translog Lock files - This is a negative case where the assertion will trip.
@@ -239,10 +238,10 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertTrue(
             result.buildAsString()
-                .startsWith(String.join(SEPARATOR, FNV_1A_BASE64.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
+                .startsWith(String.join(SEPARATOR, FNV_1A.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
         );
 
         // assert with exact value for known base path
@@ -253,7 +252,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertEquals("AUBRfCIuWdk/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/data/", result.buildAsString());
 
         // Segment Metadata
@@ -265,10 +264,10 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertTrue(
             result.buildAsString()
-                .startsWith(String.join(SEPARATOR, FNV_1A_BASE64.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
+                .startsWith(String.join(SEPARATOR, FNV_1A.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
         );
 
         // assert with exact value for known base path
@@ -279,7 +278,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertEquals("erwR-G735Uw/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/metadata/", result.buildAsString());
 
         // Segment Lockfiles
@@ -291,10 +290,10 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertTrue(
             result.buildAsString()
-                .startsWith(String.join(SEPARATOR, FNV_1A_BASE64.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
+                .startsWith(String.join(SEPARATOR, FNV_1A.hash(pathInput), basePath, dataCategory.getName(), dataType.getName()))
         );
 
         // assert with exact value for known base path
@@ -305,195 +304,8 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_PREFIX.path(pathInput, FNV_1A);
         assertEquals("KeYDIk0mJXI/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/lock_files/", result.buildAsString());
-    }
-
-    public void testGeneratePathForHashedPrefixTypeAndFNVCompositeHashAlgorithm() {
-        BlobPath blobPath = new BlobPath();
-        List<String> pathList = getPathList();
-        for (String path : pathList) {
-            blobPath = blobPath.add(path);
-        }
-
-        String indexUUID = randomAlphaOfLength(10);
-        String shardId = String.valueOf(randomInt(100));
-        DataCategory dataCategory = TRANSLOG;
-        DataType dataType = DATA;
-
-        String basePath = getPath(pathList) + indexUUID + SEPARATOR + shardId;
-        // Translog Data
-        PathInput pathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        BlobPath result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertTrue(
-            result.buildAsString()
-                .startsWith(
-                    String.join(SEPARATOR, FNV_1A_COMPOSITE_1.hash(pathInput), basePath, dataCategory.getName(), dataType.getName())
-                )
-        );
-
-        // assert with exact value for known base path
-        BlobPath fixedBlobPath = BlobPath.cleanPath().add("xjsdhj").add("ddjsha").add("yudy7sd").add("32hdhua7").add("89jdij");
-        String fixedIndexUUID = "k2ijhe877d7yuhx7";
-        String fixedShardId = "10";
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertEquals("D10000001001000/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/translog/data/", result.buildAsString());
-
-        // Translog Metadata
-        dataType = METADATA;
-        pathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertTrue(
-            result.buildAsString()
-                .startsWith(
-                    String.join(SEPARATOR, FNV_1A_COMPOSITE_1.hash(pathInput), basePath, dataCategory.getName(), dataType.getName())
-                )
-        );
-
-        // assert with exact value for known base path
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertEquals(
-            "o00101001010011/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/translog/metadata/",
-            result.buildAsString()
-        );
-
-        // Translog Lock files - This is a negative case where the assertion will trip.
-        dataType = LOCK_FILES;
-        PathInput finalPathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        assertThrows(AssertionError.class, () -> HASHED_PREFIX.path(finalPathInput, null));
-
-        // assert with exact value for known base path
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        assertThrows(AssertionError.class, () -> HASHED_PREFIX.path(finalPathInput, null));
-
-        // Segment Data
-        dataCategory = SEGMENTS;
-        dataType = DATA;
-        pathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertTrue(
-            result.buildAsString()
-                .startsWith(
-                    String.join(SEPARATOR, FNV_1A_COMPOSITE_1.hash(pathInput), basePath, dataCategory.getName(), dataType.getName())
-                )
-        );
-
-        // assert with exact value for known base path
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertEquals("A01010000000101/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/data/", result.buildAsString());
-
-        // Segment Metadata
-        dataType = METADATA;
-        pathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertTrue(
-            result.buildAsString()
-                .startsWith(
-                    String.join(SEPARATOR, FNV_1A_COMPOSITE_1.hash(pathInput), basePath, dataCategory.getName(), dataType.getName())
-                )
-        );
-
-        // assert with exact value for known base path
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertEquals(
-            "e10101111000001/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/metadata/",
-            result.buildAsString()
-        );
-
-        // Segment Lockfiles
-        dataType = LOCK_FILES;
-        pathInput = PathInput.builder()
-            .basePath(blobPath)
-            .indexUUID(indexUUID)
-            .shardId(shardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertTrue(
-            result.buildAsString()
-                .startsWith(
-                    String.join(SEPARATOR, FNV_1A_COMPOSITE_1.hash(pathInput), basePath, dataCategory.getName(), dataType.getName())
-                )
-        );
-
-        // assert with exact value for known base path
-        pathInput = PathInput.builder()
-            .basePath(fixedBlobPath)
-            .indexUUID(fixedIndexUUID)
-            .shardId(fixedShardId)
-            .dataCategory(dataCategory)
-            .dataType(dataType)
-            .build();
-        result = HASHED_PREFIX.path(pathInput, FNV_1A_COMPOSITE_1);
-        assertEquals(
-            "K01111001100000/xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/k2ijhe877d7yuhx7/10/segments/lock_files/",
-            result.buildAsString()
-        );
     }
 
     public void testGeneratePathForHashedInfixType() {
@@ -518,7 +330,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        BlobPath result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        BlobPath result = HASHED_INFIX.path(pathInput, FNV_1A);
         String expected = derivePath(basePath, pathInput);
         String actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -534,7 +346,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = "xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/DgSI70IciXs/k2ijhe877d7yuhx7/10/translog/data/";
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -549,7 +361,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataType(dataType)
             .build();
 
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = derivePath(basePath, pathInput);
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -562,7 +374,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = "xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/oKU5SjILiy4/k2ijhe877d7yuhx7/10/translog/metadata/";
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -598,7 +410,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = derivePath(basePath, pathInput);
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -611,7 +423,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = "xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/AUBRfCIuWdk/k2ijhe877d7yuhx7/10/segments/data/";
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -625,7 +437,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = derivePath(basePath, pathInput);
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -638,7 +450,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = "xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/erwR-G735Uw/k2ijhe877d7yuhx7/10/segments/metadata/";
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -652,7 +464,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = derivePath(basePath, pathInput);
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -665,7 +477,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             .dataCategory(dataCategory)
             .dataType(dataType)
             .build();
-        result = HASHED_INFIX.path(pathInput, FNV_1A_BASE64);
+        result = HASHED_INFIX.path(pathInput, FNV_1A);
         expected = "xjsdhj/ddjsha/yudy7sd/32hdhua7/89jdij/KeYDIk0mJXI/k2ijhe877d7yuhx7/10/segments/lock_files/";
         actual = result.buildAsString();
         assertTrue(new ParameterizedMessage("expected={} actual={}", expected, actual).getFormattedMessage(), actual.startsWith(expected));
@@ -675,7 +487,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
         return "".equals(basePath)
             ? String.join(
                 SEPARATOR,
-                FNV_1A_BASE64.hash(pathInput),
+                FNV_1A.hash(pathInput),
                 pathInput.indexUUID(),
                 pathInput.shardId(),
                 pathInput.dataCategory().getName(),
@@ -684,7 +496,7 @@ public class RemoteStoreEnumsTests extends OpenSearchTestCase {
             : String.join(
                 SEPARATOR,
                 basePath,
-                FNV_1A_BASE64.hash(pathInput),
+                FNV_1A.hash(pathInput),
                 pathInput.indexUUID(),
                 pathInput.shardId(),
                 pathInput.dataCategory().getName(),